requirements.txt
requests
python-dotenv 
fastapi
uvicorn 
jinja2 
python-multipart
sqlalchemy